requirements.txt
annotated-types==0.6.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
frozendict==2.4.0
groundlight==0.13.2
idna==3.6
numpy==1.24.4
opencv-python==4.9.0.80
Pillow==9.5.0
playsound==1.3.0
pydantic==2.6.3
pydantic-core==2.16.3
python-dateutil==2.9.0.post0
requests==2.31.0
six==1.16.0
typer==0.9.0
typing-extensions==4.10.0
urllib3==1.26.18
